backend/requirements.txt
flask==3.1.0
flask-cors==6.0.2
pyodbc==5.3.0
python-dotenv==1.2.1
openai==1.68.2
